ov7670/v2/python/requirements.txt
pyserial>=3.5
matplotlib>=3.5.0
numpy>=1.21.0
Pillow>=9.0.0 